requirements.txt
# Core dependencies
requests>=2.32.0
python-dotenv>=1.2.0
rich>=13.0.0

# New dependencies for grok-code
pygments>=2.17.0
glob2>=0.7
click>=8.1.0

# Testing dependencies
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
